+++ Hooks/Paper/src/main/java/fr/maxlego08/menu/hooks/dialogs/loader/body/DialogPlainMessageBodyLoader.java
@@ -7,10 +7,13 @@ import fr.maxlego08.menu.api.button.Button;
 import fr.maxlego08.menu.api.button.DefaultButtonValue;
 import fr.maxlego08.menu.api.button.buttons.dialogs.body.DialogPlainMessageBody;
 import fr.maxlego08.menu.api.loader.ButtonLoader;
+import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.jspecify.annotations.NonNull;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @AutoButtonLoader
 @RequireSupport(RequireSupport.SupportType.DIALOG)
@@ -25,6 +28,17 @@ public class DialogPlainMessageBodyLoader extends ButtonLoader {
         List<String> messages = configuration.getStringList(path+".messages");
         int width = configuration.getInt(path+".width",128);
 
-        return new DialogPlainMessageBody(messages, width);
+        Map<String, List<String>> localizedMessages = null;
+
+        ConfigurationSection localizedMessageSection = configuration.getConfigurationSection(path + ".localized-messages");
+        if (localizedMessageSection != null) {
+            localizedMessages = new HashMap<>();
+            for (String locale : localizedMessageSection.getKeys(false)) {
+                List<String> localizedMessageList = configuration.getStringList(path + ".localized-messages." + locale);
+                localizedMessages.put(locale, localizedMessageList);
+            }
+        }
+
+        return new DialogPlainMessageBody(messages, width, localizedMessages);
     }
 }