+++ src/main/java/fr/maxlego08/menu/ZMenuPlugin.java
@@ -327,6 +327,10 @@ public class ZMenuPlugin extends ZPlugin implements fr.maxlego08.menu.api.MenuPl
 //         this.inventoryManager.registerInventoryListener(this.packetUtils);
         if (this.isActive(Plugins.PACKETEVENTS)) this.inventoryManager.registerInventoryListener(new PacketEventPlayerInventoryManager(this));
 
+        if (this.packetManager != null) {
+            this.packetManager.onPostEnable();
+        }
+
         this.postEnable();
     }
 
@@ -441,6 +445,8 @@ public class ZMenuPlugin extends ZPlugin implements fr.maxlego08.menu.api.MenuPl
                 nmsMenuPacketListener.shutdown();
             }
         }
+
+        OfflinePlayerCache.uninstall(this);
 
         this.itemManager.unloadListeners();