+++ src/main/java/fr/maxlego08/menu/command/commands/players/CommandMenuPlayersRemove.java
@@ -1,12 +1,12 @@
 package fr.maxlego08.menu.command.commands.players;
 
-import com.mojang.brigadier.arguments.StringArgumentType;
 import fr.maxlego08.menu.ZMenuPlugin;
 import fr.maxlego08.menu.api.players.DataManager;
 import fr.maxlego08.menu.api.players.PlayerData;
 import fr.maxlego08.menu.api.utils.Message;
 import fr.maxlego08.menu.common.enums.Permission;
 import fr.maxlego08.menu.common.utils.MessageUtils;
+import fr.maxlego08.menu.common.utils.command.NonSpaceStringArgumentType;
 import fr.robie.paperdispatch.argument.OfflinePlayerArgument;
 import fr.robie.paperdispatch.command.CommandDispatch;
 import fr.robie.paperdispatch.command.CommandResultType;
@@ -24,7 +24,7 @@ public class CommandMenuPlayersRemove extends SubCommand<ZMenuPlugin> {
         super(plugin, "remove", "r");
         this.setPermission(Permission.ZMENU_PLAYERS_REMOVE.getPermission());
         this.addRequiredArgument("player", new OfflinePlayerArgument());
-        this.addRequiredArgument(Commands.argument("key", StringArgumentType.string()).suggests((ctx, builder) -> {
+        this.addRequiredArgument(Commands.argument("key", new NonSpaceStringArgumentType()).suggests((ctx, builder) -> {
             UUID targetId = ctx.getArgument("player", UUID.class);
             this.plugin.getDataManager().getKeys(targetId).stream().filter(entry -> entry.toLowerCase().startsWith(builder.getRemainingLowerCase()))
                     .forEach(builder::suggest);