+++ src/main/java/fr/maxlego08/menu/command/commands/players/CommandMenuPlayersRemoveAll.java
@@ -1,11 +1,11 @@
 package fr.maxlego08.menu.command.commands.players;
 
-import com.mojang.brigadier.arguments.StringArgumentType;
 import fr.maxlego08.menu.ZMenuPlugin;
 import fr.maxlego08.menu.api.players.DataManager;
 import fr.maxlego08.menu.api.utils.Message;
 import fr.maxlego08.menu.common.enums.Permission;
 import fr.maxlego08.menu.common.utils.MessageUtils;
+import fr.maxlego08.menu.common.utils.command.NonSpaceStringArgumentType;
 import fr.robie.paperdispatch.command.CommandDispatch;
 import fr.robie.paperdispatch.command.CommandResultType;
 import fr.robie.paperdispatch.command.SubCommand;
@@ -17,7 +17,7 @@ public class CommandMenuPlayersRemoveAll extends SubCommand<ZMenuPlugin> {
     public CommandMenuPlayersRemoveAll(ZMenuPlugin plugin) {
         super(plugin, "removeall");
         this.setPermission(Permission.ZMENU_PLAYERS_REMOVE_ALL.getPermission());
-        this.addRequiredArgument(Commands.argument("key", StringArgumentType.string()).suggests((ctx, builder) -> {
+        this.addRequiredArgument(Commands.argument("key", new NonSpaceStringArgumentType()).suggests((ctx, builder) -> {
             this.plugin.getDataManager().getKeys().stream().filter(entry -> entry.toLowerCase().startsWith(builder.getRemainingLowerCase()))
                     .forEach(builder::suggest);
             return builder.buildFuture();