+++ src/main/java/fr/maxlego08/menu/command/commands/players/CommandMenuPlayersSubtract.java
@@ -1,13 +1,13 @@
 package fr.maxlego08.menu.command.commands.players;
 
 import com.mojang.brigadier.arguments.IntegerArgumentType;
-import com.mojang.brigadier.arguments.StringArgumentType;
 import fr.maxlego08.menu.ZMenuPlugin;
 import fr.maxlego08.menu.api.players.Data;
 import fr.maxlego08.menu.api.players.DataManager;
 import fr.maxlego08.menu.api.utils.Message;
 import fr.maxlego08.menu.common.enums.Permission;
 import fr.maxlego08.menu.common.utils.MessageUtils;
+import fr.maxlego08.menu.common.utils.command.NonSpaceStringArgumentType;
 import fr.maxlego08.menu.players.ZData;
 import fr.robie.paperdispatch.argument.OfflinePlayerArgument;
 import fr.robie.paperdispatch.cache.OfflinePlayerCache;
@@ -26,7 +26,7 @@ public class CommandMenuPlayersSubtract extends SubCommand<ZMenuPlugin> {
         super(plugin, "subtract", "sub");
         this.setPermission(Permission.ZMENU_PLAYERS_SUBTRACT.getPermission());
         this.addRequiredArgument("player", new OfflinePlayerArgument());
-        this.addRequiredArgument(Commands.argument("key", StringArgumentType.string()).suggests((ctx, builder) -> {
+        this.addRequiredArgument(Commands.argument("key", new NonSpaceStringArgumentType()).suggests((ctx, builder) -> {
             UUID targetId = ctx.getArgument("player", UUID.class);
             this.plugin.getDataManager().getKeys(targetId).stream().filter(entry -> entry.toLowerCase().startsWith(builder.getRemainingLowerCase()))
                     .forEach(builder::suggest);